Implement block-level scope
Block statements like `if-else` and `while`, introduce a new scope.
This is slightly different from function scopes.

Function scopes inherit variables from parent scope, but can shadow them
by declaring new variables with the same name.

In a block scope, a variable from the outer scope cannot be shadowed.
Trying to declare a variable with the same name from the parent scope
will throw an error. But like functions, all variables that are
exclusively declared inside, go out-of-scope and are unavailable outside.